VCV-121: addressed review feedback
