+++ src/features/review/components/site-list/review-site-list-header.tsx
@@ -1,11 +1,5 @@
 'use client';
 
-import { Button } from '@/components/ui/button';
-import {
-    Popover,
-    PopoverContent,
-    PopoverTrigger,
-} from '@/components/ui/popover';
 import {
     Select,
     SelectContent,
@@ -15,14 +9,7 @@ import {
     SelectTrigger,
     SelectValue,
 } from '@/components/ui/select';
-import { format } from 'date-fns';
-import { CalendarIcon, ChevronLeft, ChevronRight } from 'lucide-react';
-import { useState } from 'react';
-
-const MONTHS = [
-    'Jan', 'Feb', 'Mar', 'Apr', 'May', 'Jun',
-    'Jul', 'Aug', 'Sep', 'Oct', 'Nov', 'Dec',
-];
+import MonthPicker from '@/features/review/components/site-list/month-picker';
 
 interface ReviewSiteListHeaderProps {
     districts: string[];
@@ -39,22 +26,6 @@ export default function ReviewSiteListHeader({
     selectedMonth,
     onMonthChange,
 }: ReviewSiteListHeaderProps) {
-    const [open, setOpen] = useState(false);
-    const [viewYear, setViewYear] = useState(selectedMonth.getFullYear());
-
-    function goToPreviousYear() {
-        setViewYear(year => year - 1);
-    }
-
-    function goToNextYear() {
-        setViewYear(year => year + 1);
-    }
-
-    function selectMonth(monthIndex: number) {
-        onMonthChange(new Date(viewYear, monthIndex, 1));
-        setOpen(false);
-    }
-
     return (
         <div className="flex items-center justify-between">
             <Select
@@ -76,53 +47,10 @@ export default function ReviewSiteListHeader({
                 </SelectContent>
             </Select>
 
-            <Popover open={open} onOpenChange={setOpen}>
-                <PopoverTrigger asChild>
-                    <Button variant="outline" className="w-44 justify-start gap-2">
-                        <CalendarIcon className="h-4 w-4 shrink-0" />
-                        {format(selectedMonth, 'MMMM yyyy')}
-                    </Button>
-                </PopoverTrigger>
-                <PopoverContent className="w-auto p-3" align="end">
-                    <div className="flex items-center justify-between mb-3">
-                        <Button
-                            variant="ghost"
-                            size="icon"
-                            onClick={goToPreviousYear}
-                            aria-label="Previous year"
-                        >
-                            <ChevronLeft className="h-4 w-4" />
-                        </Button>
-                        <span className="text-sm font-medium">{viewYear}</span>
-                        <Button
-                            variant="ghost"
-                            size="icon"
-                            onClick={goToNextYear}
-                            aria-label="Next year"
-                        >
-                            <ChevronRight className="h-4 w-4" />
-                        </Button>
-                    </div>
-                    <div className="grid grid-cols-3 gap-1">
-                        {MONTHS.map((month, index) => (
-                            <Button
-                                key={month}
-                                variant={
-                                    selectedMonth.getMonth() === index &&
-                                    selectedMonth.getFullYear() === viewYear
-                                        ? 'default'
-                                        : 'ghost'
-                                }
-                                size="sm"
-                                className="h-8"
-                                onClick={() => selectMonth(index)}
-                            >
-                                {month}
-                            </Button>
-                        ))}
-                    </div>
-                </PopoverContent>
-            </Popover>
+            <MonthPicker
+                selectedMonth={selectedMonth}
+                onMonthChange={onMonthChange}
+            />
         </div>
     );
 }
